refactor: update structure for auth components
