+++ common/components/auth/sign-up/button-handler.tsx
