+++ common/components/auth/sign-up/highlight-bar.tsx
