+++ common/components/auth/sign-up/registration-step.tsx
