+++ common/components/auth/sign-up/type-selection-form.tsx
