+++ common/components/auth/sign-up/user-type-card.tsx
